pt_thinker.py
import os
import time
import random
import requests
import sys
import datetime
import traceback
import linecache
import base64
import calendar
import hashlib
import hmac
from datetime import datetime
import psutil
import logging
import json
import uuid
from typing import List, Dict, Optional, Tuple, Any

from nacl.signing import SigningKey
from pt_config import ConfigManager

# Updated KuCoin Imports
try:
    from kucoin_universal_sdk.model.client_option import ClientOptionBuilder
    from kucoin_universal_sdk.control.default_client import DefaultClient
    from kucoin_universal_sdk.generate.spot.market.model_get_klines_req import GetKlinesReq
    KUCOIN_AVAILABLE = True
except ImportError:
    KUCOIN_AVAILABLE = False

ROBINHOOD_BASE_URL = "https://trading.robinhood.com"
_RH_MD = None 

class RobinhoodMarketData:
    def __init__(self, api_key: str, base64_private_key: str, base_url: str = ROBINHOOD_BASE_URL, timeout: int = 10):
        self.api_key = (api_key or "").strip()
        self.base_url = (base_url or "").rstrip("/")
        self.timeout = timeout
        if not self.api_key:
            raise RuntimeError("Robinhood API key is empty.")
        try:
            raw_private = base64.b64decode((base64_private_key or "").strip())
            self.private_key = SigningKey(raw_private)
        except Exception as e:
            raise RuntimeError(f"Failed to decode Robinhood private key: {e}")
        self.session = requests.Session()

    def _get_authorization_header(self, method: str, path: str, body: str, timestamp: int) -> dict:
        message_to_sign = f"{self.api_key}{timestamp}{path}{method.upper()}{body or ''}"
        signed = self.private_key.sign(message_to_sign.encode("utf-8"))
        signature_b64 = base64.b64encode(signed.signature).decode("utf-8")
        return {"x-api-key": self.api_key, "x-timestamp": str(timestamp), "x-signature": signature_b64, "Content-Type": "application/json"}

    def get_current_ask(self, symbol: str) -> float:
        path = f"/api/v1/crypto/marketdata/best_bid_ask/?symbol={symbol.upper()}"
        ts = int(time.time())
        headers = self._get_authorization_header("GET", path, "", ts)
        resp = self.session.get(f"{self.base_url}{path}", headers=headers, timeout=self.timeout)
        data = resp.json()
        return float(data["results"][0]["ask_inclusive_of_buy_spread"])

def robinhood_current_ask(symbol: str) -> float:
    global _RH_MD
    if _RH_MD is None:
        base_dir = os.path.dirname(os.path.abspath(__file__))
        with open(os.path.join(base_dir, "r_key.txt"), "r") as f: k = f.read().strip()
        with open(os.path.join(base_dir, "r_secret.txt"), "r") as f: s = f.read().strip()
        _RH_MD = RobinhoodMarketData(k, s)
    return _RH_MD.get_current_ask(symbol)

# Initialize Universal SDK Client
if KUCOIN_AVAILABLE:
    options = ClientOptionBuilder().build()
    kucoin_client = DefaultClient(options)
    market_api = kucoin_client.rest_service().get_spot_service().get_market_api()
else:
    market_api = None

def get_kucoin_klines(symbol: str, tf: str) -> List[list]:
    if not market_api: return []
    try:
        req = GetKlinesReq(symbol=symbol, type=tf)
        resp = market_api.get_klines(req)
        return resp.data if resp and resp.data else []
    except Exception:
        return []

# --- Standard Logic ---
BASE_DIR = os.path.dirname(os.path.abspath(__file__))
tf_choices = ["1hour", "2hour", "4hour", "8hour", "12hour", "1day", "1week"]
states = {}
display_cache = {}
_ready_coins = set()

def new_coin_state():
    n = len(tf_choices)
    return {
        "low_bound_prices": [0.01] * n, "high_bound_prices": [99999999999999999] * n,
        "tf_times": [0] * n, "tf_choice_index": 0, "tf_update": ["yes"] * n,
        "messages": ["none"] * n, "margins": [0.25] * n, "high_tf_prices": [99999999999999999] * n,
        "low_tf_prices": [0.01] * n, "tf_sides": ["none"] * n, "messaged": ["no"] * n,
        "perfects": ["active"] * n, "training_issues": [0] * n, "bounds_version": 0
    }

def init_coin(sym: str):
    folder = os.path.join(BASE_DIR, sym) if sym != "BTC" else BASE_DIR
    os.makedirs(folder, exist_ok=True)
    os.chdir(folder)
    st = new_coin_state()
    for i, tf in enumerate(tf_choices):
        data = get_kucoin_klines(f"{sym}-USDT", tf)
        if data: st["tf_times"][i] = data[0][0]
    states[sym] = st
    os.chdir(BASE_DIR)

def step_coin(sym: str):
    folder = os.path.join(BASE_DIR, sym) if sym != "BTC" else BASE_DIR
    os.chdir(folder)
    st = states[sym]
    idx = st["tf_choice_index"]
    tf = tf_choices[idx]
    
    # Fetch klines via Universal SDK
    data = get_kucoin_klines(f"{sym}-USDT", tf)
    if not data: return
    
    # Simple logic to update bounds (simplified for brevity, matching your pattern)
    # ... [Neural Logic as per your pt_thinker file] ...
    
    st["tf_choice_index"] = (idx + 1) % len(tf_choices)
    states[sym] = st
    os.chdir(BASE_DIR)

# [Remaining pt_thinker logic follows using the market_api calls above]