requirements.txt
# Update these lines in bobtrader/requirements.txt
requests
psutil
matplotlib
colorama
cryptography
PyNaCl
kucoin-universal-sdk  # Replaces kucoin-python
setuptools            # Add this as a fallback for older modules

# Notifications
yagmail
discord-webhook
python-telegram-bot

# Configuration
pyyaml

# Risk management
scipy